fix(broker): never reuse a subscriber id (#397)
A subscriber id was a slab slot, and slab hands a freed slot to the next
insert. Both unregister paths work from an id captured earlier and neither
holds a lock across the gap, so an id could come to name a different
subscriber than the caller meant:

- The publish fanout iterates a snapshot taken under the log lock and
  released before sending. A subscriber that departs during the fanout and
  is replaced before it ends leaves the reap removing the replacement.
- A subscription's receiver drops before its guard — they are separate
  fields, and into_parts hands them out separately — so a reap can free the
  registration in between and the guard's own unregister then lands on
  whoever took the slot.

Either way a subscriber that did nothing wrong is unregistered while
holding a live subscription. Its channel closes once the last snapshot
clone drops and recv() returns None, which is indistinguishable from the
broker closing the subscription.

Reaping by identity would close the first path; it would not close the
second, because the guard holds no sender to compare. Making ids monotonic
closes both, so the registry is a HashMap keyed by a counter. The slab
bought dense indexing the fanout never used — it reads the snapshot, which
is rebuilt on every register and remove either way — so the only cost is
sorting the snapshot to keep fanout order stable across publishes.

Reintroducing reuse fails all three of the new cases.

Fixes #256

+++ crates/felix-broker/src/broker.rs
@@ -522,7 +522,7 @@ impl Broker {
                         ));
                         sent += item_count;
                     } else {
-                        closed_subscribers.push(subscriber.id as u64);
+                        closed_subscribers.push(subscriber.id);
                     }
                 }
                 SubQueuePolicy::DropNew | SubQueuePolicy::DropOld => {
@@ -550,7 +550,7 @@ impl Broker {
                             }
                         }
                         Err(tokio::sync::mpsc::error::TrySendError::Closed(_)) => {
-                            closed_subscribers.push(subscriber.id as u64);
+                            closed_subscribers.push(subscriber.id);
                         }
                     }
                 }
@@ -826,7 +826,7 @@ impl Broker {
         // Built the moment the subscriber exists, so every error path below
         // releases the registration by `Drop` instead of stranding a closed
         // sender in the registry for the publish path to reap later. Repeated
-        // rejected subscribes would otherwise grow the slab without ever
+        // rejected subscribes would otherwise grow the registry without ever
         // touching the per-connection subscription cap.
         let guard = SubscriptionGuard {
             stream_state: Arc::downgrade(&stream_state),
